Added FadeIn component
