+++ pages/components/Projects/DM.js
@@ -1,27 +1,22 @@
-import { motion } from "framer-motion";
+import FadeIn from "../FadeIn";
 import Link from "next/link";
-import Chip from '@material-ui/core/Chip';
+import Chip from "@material-ui/core/Chip";
 
 export default function DM() {
-  const easing = [0.6, -0.05, 0.01, 0.99];
-  const fadeIn = {
-    initial: {
-      opacity: 0,
-    },
-    animate: {
-      opacity: 1,
-      transition: {
-        duration: 1,
-        ease: easing,
-      },
-    },
-  };
-
-  const tech = ["Next.js", "React", "PHP", "Axios", "Bootstrap", "MariaDB", "AJAX", "jQuery"];
+  const tech = [
+    "Next.js",
+    "React",
+    "PHP",
+    "Axios",
+    "Bootstrap",
+    "MariaDB",
+    "AJAX",
+    "jQuery",
+  ];
 
   return (
-    <motion.div exit="exit" initial="initial" animate="animate">
-      <motion.div variants={fadeIn}>
+    <>
+      <FadeIn>
         <hr className="mt-4" />
         <div>
           <div>
@@ -30,7 +25,7 @@ export default function DM() {
           <p className="mt-2">
             Bunch of personal websites and developer portfolio sites for myself,
             ranging from the maximal to super minimal design philosophy.
-            </p>
+          </p>
           {/* <h6 className="mt-4">Releases</h6> */}
           <div className="d-none">
             <ul className="nav">
@@ -210,13 +205,11 @@ export default function DM() {
           <h6 className="mt-4">Technologies</h6>
           <div class="tech-badges">
             {tech.map((chip) => {
-              return (
-                <Chip size="small" key={chip} label={chip} />
-              );
+              return <Chip size="small" key={chip} label={chip} />;
             })}
           </div>
         </div>
-      </motion.div>
-    </motion.div>
+      </FadeIn>
+    </>
   );
 }
