+++ pages/components/Projects/JailBird.js
@@ -1,27 +1,28 @@
-import { motion } from "framer-motion";
+import FadeIn from "../FadeIn";
 import Link from "next/link";
-import Chip from '@material-ui/core/Chip';
+import Chip from "@material-ui/core/Chip";
 
 export default function JailBird() {
-  const easing = [0.6, -0.05, 0.01, 0.99];
-  const fadeIn = {
-    initial: {
-      opacity: 0,
-    },
-    animate: {
-      opacity: 1,
-      transition: {
-        duration: 1.5,
-        ease: easing,
-      },
-    },
-  };
-
-  const tech = ["PHP", "Apache", "Nginx", "Shell", "Bootstrap", "MariaDB", "MongoDB", "Firestore", "AJAX", "jQuery", "Linux", "Git", "Cloudflare", "Vercel"];
+  const tech = [
+    "PHP",
+    "Apache",
+    "Nginx",
+    "Shell",
+    "Bootstrap",
+    "MariaDB",
+    "MongoDB",
+    "Firestore",
+    "AJAX",
+    "jQuery",
+    "Linux",
+    "Git",
+    "Cloudflare",
+    "Vercel",
+  ];
 
   return (
-    <motion.div exit="exit" initial="initial" animate="animate">
-      <motion.div variants={fadeIn}>
+    <>
+      <FadeIn>
         <hr className="mt-4" />
         <div>
           <div>
@@ -43,13 +44,11 @@ export default function JailBird() {
           <h6 className="mt-4">Technologies</h6>
           <div class="tech-badges">
             {tech.map((chip) => {
-              return (
-                <Chip size="small" key={chip} label={chip} />
-              );
+              return <Chip size="small" key={chip} label={chip} />;
             })}
           </div>
         </div>
-      </motion.div>
-    </motion.div>
+      </FadeIn>
+    </>
   );
 }
