+++ pages/components/Projects/KFCEcom.js
@@ -1,39 +1,34 @@
 import React from "react";
-import { motion } from "framer-motion";
+import FadeIn from "../FadeIn";
 import Link from "next/link";
-import Chip from '@material-ui/core/Chip';
+import Chip from "@material-ui/core/Chip";
 
 export default function KFCEcom() {
-  const easing = [0.6, -0.05, 0.01, 0.99];
-  const fadeIn = {
-    initial: {
-      opacity: 0,
-    },
-    animate: {
-      opacity: 1,
-      transition: {
-        duration: 1.5,
-        ease: easing,
-      },
-    },
-  };
-
-  const tech = ["React", "PHP", "Axios", "Bootstrap", "MariaDB", "jQuery", "Cloudflare", "Vercel"];
+  const tech = [
+    "React",
+    "PHP",
+    "Axios",
+    "Bootstrap",
+    "MariaDB",
+    "jQuery",
+    "Cloudflare",
+    "Vercel",
+  ];
 
   return (
-    <motion.div exit="exit" initial="initial" animate="animate">
-      <motion.div variants={fadeIn}>
+    <>
+      <FadeIn>
         <hr className="mt-4" />
         <div>
           <div>
             <h6 className="middle-underline d-inline">
               KuTumba FC E-commerce Platform
-          </h6>
+            </h6>
           </div>
           <p className="mt-2">
-            E-commerce platform with monthly recurring payments made possible by a
-            Razorpay integration to sell merchandise for KuTumba FC.
-        </p>
+            E-commerce platform with monthly recurring payments made possible by
+            a Razorpay integration to sell merchandise for KuTumba FC.
+          </p>
           {/* <h6 className="text-grey mt-4">Releases</h6> */}
           <div className="d-none">
             <Link passHref href={"https://v5.thedanielmark.com/"}>
@@ -43,13 +38,11 @@ export default function KFCEcom() {
           <h6 className="text-grey mt-4">Technologies</h6>
           <div class="tech-badges">
             {tech.map((chip) => {
-              return (
-                <Chip size="small" key={chip} label={chip} />
-              );
+              return <Chip size="small" key={chip} label={chip} />;
             })}
           </div>
         </div>
-      </motion.div>
-    </motion.div>
+      </FadeIn>
+    </>
   );
 }
