+++ pages/components/Projects/NowMyWatchBegins.js
@@ -1,35 +1,31 @@
-import { motion } from "framer-motion";
+import FadeIn from "../FadeIn";
 import Link from "next/link";
-import Chip from '@material-ui/core/Chip';
+import Chip from "@material-ui/core/Chip";
 
 export default function NowMyWatchBegins() {
-  const easing = [0.6, -0.05, 0.01, 0.99];
-  const fadeIn = {
-    initial: {
-      opacity: 0,
-    },
-    animate: {
-      opacity: 1,
-      transition: {
-        duration: 1.5,
-        ease: easing,
-      },
-    },
-  };
-
-  const tech = ["PHP", "Python", "Apache", "Bootstrap", "MariaDB", "AJAX", "Git", "Cloudflare"];
+  const tech = [
+    "PHP",
+    "Python",
+    "Apache",
+    "Bootstrap",
+    "MariaDB",
+    "AJAX",
+    "Git",
+    "Cloudflare",
+  ];
 
   return (
-    <motion.div exit="exit" initial="initial" animate="animate">
-      <motion.div variants={fadeIn}>
+    <>
+      <FadeIn>
         <hr className="mt-4" />
         <div>
           <div>
             <h6 className="middle-underline d-inline">Now My Watch Begins</h6>
           </div>
           <p className="mt-2">
-            Website where the first person to crack a complex program and enter the code
-            and his/her name would have their name appear on the winner screen.
+            Website where the first person to crack a complex program and enter
+            the code and his/her name would have their name appear on the winner
+            screen.
           </p>
           {/* <h6 className="mt-4">Releases</h6> */}
           <div className="d-none">
@@ -40,13 +36,11 @@ export default function NowMyWatchBegins() {
           <h6 className="mt-4">Technologies</h6>
           <div class="tech-badges">
             {tech.map((chip) => {
-              return (
-                <Chip size="small" key={chip} label={chip} />
-              );
+              return <Chip size="small" key={chip} label={chip} />;
             })}
           </div>
         </div>
-      </motion.div>
-    </motion.div>
+      </FadeIn>
+    </>
   );
 }
