+++ pages/components/Projects/TheArchitectAkash.js
@@ -1,27 +1,23 @@
-import { motion } from "framer-motion";
+import FadeIn from "../FadeIn";
 import Link from "next/link";
-import Chip from '@material-ui/core/Chip';
+import Chip from "@material-ui/core/Chip";
 
 export default function TheArchitectAkash() {
-  const easing = [0.6, -0.05, 0.01, 0.99];
-  const fadeIn = {
-    initial: {
-      opacity: 0,
-    },
-    animate: {
-      opacity: 1,
-      transition: {
-        duration: 1.5,
-        ease: easing,
-      },
-    },
-  };
-
-  const tech = ["React", "PHP", "Bootstrap", "MariaDB", "AJAX", "jQuery", "Git", "Cloudflare", "Vercel"];
+  const tech = [
+    "React",
+    "PHP",
+    "Bootstrap",
+    "MariaDB",
+    "AJAX",
+    "jQuery",
+    "Git",
+    "Cloudflare",
+    "Vercel",
+  ];
 
   return (
-    <motion.div exit="exit" initial="initial" animate="animate">
-      <motion.div variants={fadeIn}>
+    <>
+      <FadeIn>
         <hr className="mt-4" />
         <div>
           <div>
@@ -42,13 +38,11 @@ export default function TheArchitectAkash() {
           <h6 className="mt-4">Technologies</h6>
           <div class="tech-badges">
             {tech.map((chip) => {
-              return (
-                <Chip size="small" key={chip} label={chip} />
-              );
+              return <Chip size="small" key={chip} label={chip} />;
             })}
           </div>
         </div>
-      </motion.div>
-    </motion.div>
+      </FadeIn>
+    </>
   );
 }
