+++ pages/components/Projects/XStackPlatform.js
@@ -1,27 +1,34 @@
-import { motion } from "framer-motion";
+import FadeIn from "../FadeIn";
 import Link from "next/link";
-import Chip from '@material-ui/core/Chip';
+import Chip from "@material-ui/core/Chip";
 
 export default function XStackPlatform() {
-  const easing = [0.6, -0.05, 0.01, 0.99];
-  const fadeIn = {
-    initial: {
-      opacity: 0,
-    },
-    animate: {
-      opacity: 1,
-      transition: {
-        duration: 1.5,
-        ease: easing,
-      },
-    },
-  };
-
-  const tech = ["PHP", "Python", "Dart", "Flutter", "Apache", "Nginx", "Shell", "Node.js", "Bootstrap", "MariaDB", "MongoDB", "Firestore", "AJAX", "jQuery", "Docker", "Linux", "Git", "OpenStack", "Cloudflare", "Vercel"];
+  const tech = [
+    "PHP",
+    "Python",
+    "Dart",
+    "Flutter",
+    "Apache",
+    "Nginx",
+    "Shell",
+    "Node.js",
+    "Bootstrap",
+    "MariaDB",
+    "MongoDB",
+    "Firestore",
+    "AJAX",
+    "jQuery",
+    "Docker",
+    "Linux",
+    "Git",
+    "OpenStack",
+    "Cloudflare",
+    "Vercel",
+  ];
 
   return (
-    <motion.div exit="exit" initial="initial" animate="animate">
-      <motion.div variants={fadeIn}>
+    <>
+      <FadeIn>
         <hr className="mt-4" />
         <div>
           <div>
@@ -72,13 +79,11 @@ export default function XStackPlatform() {
           <h6 className="mt-4">Technologies</h6>
           <div class="tech-badges">
             {tech.map((chip) => {
-              return (
-                <Chip size="small" key={chip} label={chip} />
-              );
+              return <Chip size="small" key={chip} label={chip} />;
             })}
           </div>
         </div>
-      </motion.div>
-    </motion.div>
+      </FadeIn>
+    </>
   );
 }
